Add game details to homepage without using a dialog
