+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/ManageSection/index.js
@@ -9,7 +9,7 @@ import GamesList from '../../../../GameDashboard/GamesList';
 import { type Game } from '../../../../Utils/GDevelopServices/Game';
 import PlaceholderError from '../../../../UI/PlaceholderError';
 import PlaceholderLoader from '../../../../UI/PlaceholderLoader';
-import { Column, Line } from '../../../../UI/Grid';
+import { Column, LargeSpacer, Line } from '../../../../UI/Grid';
 import Paper from '../../../../UI/Paper';
 import BackgroundText from '../../../../UI/BackgroundText';
 import {
@@ -21,12 +21,19 @@ import FlatButton from '../../../../UI/FlatButton';
 import Link from '../../../../UI/Link';
 import Window from '../../../../Utils/Window';
 import { getHelpLink } from '../../../../Utils/HelpLink';
+import GameDetails, {
+  gameDetailsTabs,
+  type GameDetailsTab,
+} from '../../../../GameDashboard/GameDetails';
+import { Tabs } from '../../../../UI/Tabs';
+import Text from '../../../../UI/Text';
 
 const publishingWikiArticle = getHelpLink('/publishing/');
 
 const styles = {
   backgroundMessage: { padding: 16 },
   buttonContainer: { minWidth: 150 },
+  gameDetailsContainer: { padding: 8, flex: 1, display: 'flex' },
 };
 
 type Props = {|
@@ -35,6 +42,10 @@ type Props = {|
   onRefreshGames: () => Promise<void>,
   onGameUpdated: Game => void,
   gamesFetchingError: ?Error,
+  openedGame: ?Game,
+  setOpenedGame: (?Game) => void,
+  currentTab: GameDetailsTab,
+  setCurrentTab: GameDetailsTab => void,
 |};
 
 const ManageSection = ({
@@ -43,6 +54,10 @@ const ManageSection = ({
   onRefreshGames,
   onGameUpdated,
   gamesFetchingError,
+  openedGame,
+  setOpenedGame,
+  currentTab,
+  setCurrentTab,
 }: Props) => {
   const authenticatedUser = React.useContext(AuthenticatedUserContext);
   const {
@@ -50,6 +65,56 @@ const ManageSection = ({
     onOpenCreateAccountDialog,
     onOpenLoginDialog,
   } = authenticatedUser;
+
+  const onBack = React.useCallback(
+    () => {
+      setCurrentTab('details');
+      setOpenedGame(null);
+    },
+    [setCurrentTab, setOpenedGame]
+  );
+
+  if (openedGame) {
+    return (
+      <SectionContainer
+        flexBody
+        title={null} // Use a smaller title below
+        backAction={onBack}
+      >
+          <Text size="title" allowSelection>
+            {openedGame.gameName}
+          </Text>
+          <LargeSpacer />
+        <SectionRow expand>
+          <Paper
+            background="medium"
+            square={false}
+            style={styles.gameDetailsContainer}
+          >
+            <ColumnStackLayout noMargin expand>
+              <Tabs
+                value={currentTab}
+                onChange={setCurrentTab}
+                options={gameDetailsTabs}
+              />
+              <GameDetails
+                game={openedGame}
+                project={project}
+                onGameUpdated={onGameUpdated}
+                onGameDeleted={() => {
+                  onBack();
+                  onRefreshGames();
+                }}
+                onLoading={() => {}}
+                currentTab={currentTab}
+                setCurrentTab={setCurrentTab}
+              />
+            </ColumnStackLayout>
+          </Paper>
+        </SectionRow>
+      </SectionContainer>
+    );
+  }
 
   return (
     <SectionContainer flexBody title={<Trans>Manage Games</Trans>}>
@@ -142,6 +207,7 @@ const ManageSection = ({
               games={games}
               onRefreshGames={onRefreshGames}
               onGameUpdated={onGameUpdated}
+              onOpenGame={setOpenedGame}
             />
           )
         ) : gamesFetchingError ? (
